Check in Bazel PGP information
Summary:
This includes the full public key for the Bazel team (48457EE0) as well
as the ownertrust entry to mark the key as ultimately trusted within a
keyring.

The contents are included directly into the `bazelisk.py` file so that
the script can be self-contained; we don’t have to worry about bundling
it with any external assets.

The ownertrust format was created by running importing the Bazel key
into a clean keyring, running `gpg --edit-key 48457EE0` and setting the
trust level to “ultimate”, then invoking `gpg --export-ownertrust`.

Test Plan:
Download <https://bazel.build/bazel-release.pub.gpg> and verify that it
exactly matches the blob checked into this commit:

    $ curl --silent 'https://bazel.build/bazel-release.pub.gpg' | shasum -a 256
    30af2ca7abfb65987cd61802ca6e352aadc6129dfb5bfc9c81f16617bc3a4416  -
    $ sed -n '/BEGIN PGP/,/END PGP/p' bazelisk.py | shasum -a 256
    30af2ca7abfb65987cd61802ca6e352aadc6129dfb5bfc9c81f16617bc3a4416  -

wchargin-branch: bazel-pgp-info