DO NOT SUBMIT: debugging Buildkite failures
Summary:
The Linux tests passed, but the macOS tests failed with no useful
output, and the Windows tests failed with an unexpected GPG invocation
error with a strange path. This commit adds some print-statements in an
attempt to investigate.

Test Plan:
Running `./bazelisk version` and `bazel test :bazelisk_test` both still
work on Python 2 and 3 on Linux.

wchargin-branch: authenticate-binaries