Fix #61: Do not execute tools/bazel when BAZELISK_SKIP_WRAPPER is set.
Bazelisk will set this environment variable when executing Bazel (or
a wrapper found in <workspace_root>/tools/bazel). This prevents Bazelisk
from executing itself in an infinite loop, in case the wrapper is
Bazelisk or somehow executes Bazelisk.

If you're using a wrapper in <workspace_root>/tools/bazel that downloads
Bazelisk and delegates to it, please set the BAZELISK_SKIP_WRAPPER
environment variable to ensure that Bazelisk directly executes Bazel.