Allow using local Bazel binaries. (#93)
Bazel versions (e.g. in USE_BAZEL_VERSION or .bazelversion) can now
refer to absolute paths on the filesystem. As an added convenience, a
tilde prefix is expanded to the user's home directory.

Example:

```sh
$ USE_BAZEL_VERSION="~/bin/bazel-1.0" bazelisk
```

This would tell Bazelisk to not download any binaries and instead
just directly use the Bazel binary in `$HOME/bin/bazel-1.0`.