Retrieve Bazel versions from GCS if GitHub is down (#132)
This fallback strategy is NOT executed when users specify a fork of Bazel. In this case the version resolution algorithm fails immediately.

Note: This change only affects the origin of Bazel's version history. In both cases the Bazel binaries are downloaded from GCS, which has already been the case before this change.

Fixes #87