De-duplicate repo selection logic
The first version of the PR determined which repo should be used (e.g. release vs candidate repo) in two places:
Once in bazelisk.go when the version had to be resolved, and a second time in the core module when a binary had to be downloaded.

This has now changed: core.Repositories got the new ResolveVersion() function that resolves a version label and also returns the appropriate download function.
As a result, bazelisk.go is no longer in the business of caring about versions and can simply call the download function (unless BAZELISK_BASE_URL is set, of course).

Positive side-effects:
- All version parsing / analysis was moved into the versions module.
- A lot of code from bazelisk.go could be removed.
- LastGreenRepo was renamed to CommitRepo and now supports commit hashes, too.