No longer get incompatible flags from GitHub
We're turning down the tracking issues for incompatible flags. As a result, `bazelisk --migrate` should simply test all incompatible flags that are present in the Bazel binary.

Please note that the new implementation is as technically incorrect as the previous one, given that it simply tests a list of incompatible flags without making sure that they're implemented for the actual Bazel command that is being executed.
However, in reality this hasn't been a problem since we started using --migrate on Bazel CI.