Consider wrapper when only user- or group-executable
The bazel wrapper might not be executable by `other`, only
`user` or `group`, depending for instance on the umask of the
user.

Allow for this by checking the user and group executable bits
too. If the file is not executable by the current user, the
run will fail with permission denied.