Suppress Apple Silicon fallback warning for commits.
Bazelisk on M1 always printed "Fallback to x86_64 because arm64 is not supported on Apple Silicon until 4.1.0" when asking for a Bazel binary built at a commit.
However, the code downloaded the ARM64 binary regardless (which was the correct behavior given that commits are very likely newer than Bazel 4.1).

This commit suppresses the warning message in this case.

Fixes https://github.com/bazelbuild/bazelisk/issues/391