Implement support for BAZELISK_VERIFY_SHA256 (#441)
The new BAZELISK_VERIFY_SHA256 variable can be set to the expected SHA256
hash of the downloaded Bazel binary.  If set, then the binary is required
to match the hash before it is used.

This is important for cases where provenance of the artifact cannot be
asserted purely via the HTTPS trust chain (such as what happens in a
mutable artifact repository with lax access controls).