Update README.md
Is this right? As someone told by a different repo that "you need bazel", and then told by the bazel docs that "you need bazelisk" — without any context other than that bazelisk will install bezel for me — I had no idea what bazelisk is or how it's meant to be used. Adding this section to the Readme would have helped me, even if it sounds terribly obvious to you.