Improve signal handling.
Every signal specially recognized by Bazel that doesn't cause an immediate exit
shouldn't cause Bazelisk to exit either, as this risks orphaning the Bazel
process and making it uncontrollable by the terminal. (You can see this in action
by sending a Ctrl-\ to Bazelisk, which causes it to immediately exit and makes it
impossible to abort the still running Bazel invocation with Ctrl-C.)

Note that forwarding is not necessary for a signal delivered by the terminal
(i.e. through Ctrl-C or Ctrl-\) because the terminal already delivers it to the
entire process group. For manually delivered signals, users should arguably
know what they're doing, so let's not go out of our way to forward them.

This also disables the printing of a Go stack dump upon SIGQUIT, which is
unhelpful: users tend to report it in place of the far more useful Java thread
dump.